+++ server/empire/src/main/java/com/example/empire/repository/UtilizatorRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 public interface UtilizatorRepository extends JpaRepository<Utilizator, String> {
     public Optional<Utilizator> getAllByUsername(String username);
-    public ArrayList<Utilizator> getAllByIdJoc(int idJoc);
+    public ArrayList<Utilizator> getAllByIdJoc(Long idJoc);
 }
